+++ packages/sns/src/governance.canister.spec.ts
@@ -2,8 +2,9 @@ import type { ActorSubclass } from "@dfinity/agent";
 import { mock } from "jest-mock-extended";
 import type { _SERVICE as SnsGovernanceService } from "../candid/sns_governance";
 import { MAX_LIST_NEURONS_RESULTS } from "./constants/governance.constants";
+import { SnsGovernanceError } from "./errors/governance.errors";
 import { SnsGovernanceCanister } from "./governance.canister";
-import { neuronsMock } from "./mocks/governance.mock";
+import { neuronIdMock, neuronMock, neuronsMock } from "./mocks/governance.mock";
 import { rootCanisterIdMock } from "./mocks/sns.mock";
 
 describe("Governance canister", () => {
@@ -34,6 +35,43 @@ describe("Governance canister", () => {
       limit: MAX_LIST_NEURONS_RESULTS,
       of_principal: [],
       start_page_at: [],
+    });
+  });
+
+  describe("getNeuron", () => {
+    it("should return the neuron", async () => {
+      const service = mock<ActorSubclass<SnsGovernanceService>>();
+      service.get_neuron.mockResolvedValue({
+        result: [{ Neuron: neuronMock }],
+      });
+
+      const canister = SnsGovernanceCanister.create({
+        canisterId: rootCanisterIdMock,
+        certifiedServiceOverride: service,
+      });
+      const res = await canister.getNeuron({
+        neuronId: neuronIdMock,
+        certified: true,
+      });
+      expect(res).toEqual(neuronMock);
+    });
+
+    it("should raise error on governance error", async () => {
+      const service = mock<ActorSubclass<SnsGovernanceService>>();
+      service.get_neuron.mockResolvedValue({
+        result: [{ Error: { error_message: "error", error_type: 2 } }],
+      });
+
+      const canister = SnsGovernanceCanister.create({
+        canisterId: rootCanisterIdMock,
+        certifiedServiceOverride: service,
+      });
+      const call = () =>
+        canister.getNeuron({
+          neuronId: neuronIdMock,
+          certified: true,
+        });
+      expect(call).rejects.toThrowError(SnsGovernanceError);
     });
   });
 });